HHH-16008 promote joinTransaction() and isJoinedToTransaction() to SharedSessionContract
so they are available on a StatelessSession
